fix: postDetail 컴포넌트 분리
