+++ src/pages/PostDetailPage/PostDetailPage.tsx
@@ -1,0 +1,1 @@
+import React from "react";
